[#164035593] Improved Scene editing
 - allow multiple entity selection.
 - Gizmo now supports manipulation with several object in the same time.
 Though non-uniform scaling does not work exactly right, while it's
 pretty strange use case at the same time.
 - copy, paste and delete is also working with multiple selection.
 - added ability to add other types of entities to the scene like:
   - lights.
   - cameras.
   - characters.
 - display pictogramms for invisible objects (lights).
 - display grid (to be replaced with proper manually drawn grid).
 - implemented entity outline on selection.

Other Changes
-------------

 - added compositor support for Ogre 1.9. Compositors can now be applied
 using `compositors` field of the viewport definition.
 - fixed billboard bugs that were causing crashes.
 - fixed crashes on object deletion (Gizmo reference cleanup problem).
 - added support for setting `OgreObject` visibility flags.
 - using more optimal way to invalidate `OgreObject` dimensions cache.
 - enabled raycasting to billboards.
 - added SSAO shaders, not enabling it as it requires OGRE compositor
 builder to be implemented (#166380554).
 - fixed some part of Windows build warnings.
 - using `const char*` instead of `std::string` for event types.
 - fixed locale bugs. Was parsing "." incorrectly for floating numbers
 encoded in text files.
 - proper HighDPI support on all platforms.
 - fixed text asset create type selector in editor (##166351970).
 - fixed nasty bugs with script paths setup during project opening.

+++ GsageCore/include/ThreadSafeQueue.h
@@ -84,10 +84,10 @@ namespace Gsage {
        * Get queue item
        * @param dest to write item to
        */
-      int get(C& dest)
+      size_t get(C& dest)
       {
         mMutex.lock();
-        int tail = mQueue.size();
+        size_t tail = mQueue.size();
         if(tail > 0) {
           dest = mQueue.front();
           mQueue.pop();
